requirements.txt
openai
flask
redis
langchain
langchain_openai
langchain_community
langchain_core
python-dotenv